+++ CHANGELOG.md
@@ -17,6 +17,10 @@
   - Markers of lightness are only merged when labelLevel == 0.
     - When merge happens, the marker shows the middle value of lightness.
 - Remove rare fantom empty marker space on scrollable legend.
+- Same legend with different measure series are not interpolated. 
+- Different legend with same dimension series are interpolated.
+- Invisible axis labels are not draw.
+- Fix chaotic axis labels on sorted chart with multiple dimension.
 
 ## [0.15.0] - 2024-10-28